More Information and Images
I finnaly implemented more information about the pokemon as awell as another scroll view in informatino view. I was able to make it so that additional information was viewable. Though, this way does make it a bit harder to access since i have to hardcode everything.

Its possible that this will make it so that there is greater memory usage, which we will need to optimize.